Split more big files
